services/auth/pyproject.toml
[project]
name = "auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "bcrypt>=5.0.0",
    "flake8>=7.3.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-jwt-extended>=4.7.1",
    "flask-limiter>=4.0.0",
    "isort>=7.0.0",
    "marshmallow>=4.0.1",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.1.1",
    "supabase>=2.22.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]
